Remove enum as-u32 cast codegen: rely on libperl-sys derive(PartialEq, PartialOrd)
libperl-sys side now sets bindgen derive_partialeq/derive_partialord,
so svtype and other enums support comparison operators natively.
Removed is_enum_expr() and enum comparison cast logic from both
macro and inline codegen paths. Reverted regression test expectations.
Updated samples/bindings.rs with new derive attributes.

+++ tests/expected_rust/Perl_CvDEPTH.rs
@@ -3,10 +3,7 @@
 pub unsafe fn Perl_CvDEPTH(sv: *const CV) -> *mut I32 {
     unsafe {
         assert!(!sv.is_null());
-        assert!(
-            (((SvTYPE(sv) as u32) == (SVt_PVCV as u32))
-                || ((SvTYPE(sv) as u32) == (SVt_PVFM as u32)))
-        );
+        assert!(((SvTYPE(sv) == SVt_PVCV) || (SvTYPE(sv) == SVt_PVFM)));
         return (&mut (*((*sv).sv_any as *mut XPVCV)).xcv_depth);
     }
 }
